Process the host list (if provided) before building objects
The host list is now processed in the main area, before objects
have been created.
If a comma separated list is provided, then a list is produced:

hosts:
  - host1
  - host2
  - host3

but if a single regex is passed via the placement option (-p),
the "host_pattern" section is created within the placement area.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>