Host list and host_pattern are now separated options
The placement class is built providing (as attributes)
the way the placement section should be rendered in the
spec.
This commit adds two different parameters to make sure
we can properly discern when a host_pattern is provided
instead of a host list.
Tests are also refactored a bit (but it's under heavy
development).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>