Add validation on the spec dict
The spec dict is provided by the operator as a cli specific
parameter.
In order to make sure only valid keys are passed, an array
containing a "per daemon" spec keys definition is provided.
The validate fuction, a key part of the make_daemon_spec
function, is now responsible to yield or not a spec section
for a given daemon.

In the current implementation no failures are propagated on
the main function, but it's something in TODO list.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>