Make the ceph_spec module able to apply the spec
When a spec is rendered/generated in a given path, a new
'apply' option is added to make it able to run the ceph
orch command and apply the daemon against an existing
ceph cluster.
If apply is false, nothing changes and the daemon is
rendered as an output dictionary. Later, using the ansible
to_nice_yaml filter, the daemon can be manipulated.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>