Add networks=[] parameter to the CephDaemonSpec
This change introduces a new option on the CephDaemonSpec definition.
Now you can specify a list of networks that should be rendered using
the related top level field.
This cannot be logically considered as an extra parameter, because
except for the monitors and osds (they have the cluster and public
networks defined, all the other daemons should provide this field).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>