Distribute admin keyring to all the Ceph _admin nodes
The nodes having an _admin label should be able to interact
with the ceph cluster using the admin keyring. This label is
typically added to the mon group, but they need the admin
keyring to run any ceph command (via cephadm shell or any
other client tool).
This change adds a new block to make sure both $cluster.conf
and the related client.admin keyring are distributed to the
relevant nodes.
As per [1] cephadm at the moment doesn't support a customized
cluster name, therefore this change add the flexibility to
handle the conf/key distribution via ansible until that will
be supported by the new tool.

[1] https://github.com/ceph/ceph/blob/master/doc/cephadm/operations.rst#client-keyrings-and-configs

Signed-off-by: Francesco Pantano <fpantano@redhat.com>