Add the ceph_patch_inventory ansible module
This patch just adds the new ceph_patch_inventory ansible module, that
is supposed to help removing (on the fly) a group from the inventory
during the FFU from wallaby to xena. In particular the inventory should
be patched when the 'cephadm_adopt' playbook (provided by ceph-ansible)
is executed, with the purpose of skipping the 'ceph_nfs' (aka 'nfss')
group.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>