Host spec now includes ceph crush location
Using a cephadm spec is now possible to define a "location" identifier
which will instruct cephadm to create a new CRUSH host located in the
specified hierachy.
This dict is not part of any daemon specification, but it's used by a
CephHostSpec and accept a dictionary like the following:

        {'root': 'default-1', 'rack': 'r1', 'host': 'h1'}

A few tests are also created to make sure we don't break the existing
use cases where a crush hierarchy is not defined.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>