Rework my multinode cephadm_deploy setup
This patch represents a rework of my multinode Ceph setup
calling TripleO w/o TripleO.
In addition to several fixes, this patch brings (and it's
supposed to test) a new feature: it's now possible to build
a customized crush hierarchy associated to the OSD nodes.
The ceph_spec library now supports the 'location' parameter,
which is a dict containing the OSD position within the crush
hierarchy.
The ceph_bootstrap ansible module is also updated to accept
the new ceph_crush_hierarchy parameter, which is supposed to
be passed as extra_var to the cephadm playbook.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>